refactor : useLoading custom hook 생성
